Don't require ActiveSupport >= 5 when building.
ActiveSupport 5.0.0.1 requires Ruby 2.2.2 or greater, but I'd like to
support Ruby versions that are less than this. AFAIK, there's nothing
actually requiring a later ActiveSupport version.

This change should ensure that the build uses a version of ActiveSupport
which is compatible with all Rubies I want to test against, without
actually constraining the version of ActiveSupport that gets used when
the gem is installed (since those constraints come only from the
`gemspec` file, not the `Gemfile`).