Support `chat_update` calls for Slack messages
This makes it simpler to update existing messages from our bots.

As part of this, the `call` method on bot has changed semantics;
previously it would call a method on the underlying Slack client, but
now it calls a method on the bot instance itself. The old behaviour is
now available with `slack_call`.