Allow bot subclasses to store details of message for updating
I've called this concept a 'pingback', although that makes it sound
grander than it is. Basically, we need to give code in the client
application the ability to handle some details of the posted
messages (the channel and `ts` timestamp, specifically), so they can
then update those messages if they want.

I don't want to couple this library to any particular way of
persisting that data, however, so instead it's a method that bot
subclasses can define to store the data in a database, for example. My
idea is that the `pingback` option given to `say` will be some
identifier which the application can then use to make a concrete
association to a real model. So, for example, it could be a GlobalID
string.